# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "whisper-wrap engine API",
    "description": "HTTP/WebSocket/SSE API for the whisper-wrap transcription engine.",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "3.0.0-beta.1"
  },
  "paths": {
...
    "/models": {
      "get": {
        "tags": [
          "models"
        ],
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Installed + registered ASR models with active/loaded state (ad-hoc JSON)."
          },
          "500": {
            "description": "Registry read error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/transcribe": {
      "post": {
        "tags": [
          "transcription"
        ],
        "operationId": "transcribe",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "prompt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "description": "Optional per-request ASR model (per-request-asr-model). Absent selects\nthe active engine; present selects that model with no global swap.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "log",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "description": "Audio payload — raw bytes with an `audio/*` (or `application/octet-stream`) Content-Type, or `multipart/form-data` with a `file` part.",
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transcription result — text plus timing/metadata (ad-hoc JSON object)."
          },
          "400": {
            "description": "Empty or unreadable audio body, or missing multipart `file` field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "Audio exceeds the configured maximum file size.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Content-Type or media format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Audio decode or inference failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "No ASR model is loaded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/audio/transcriptions": {
      "post": {
        "tags": [
          "openai-compat"
        ],
        "operationId": "transcriptions",
        "requestBody": {
          "description": "OpenAI-compatible transcription request — `multipart/form-data` with a `file` part (audio) plus optional `model`/`language`/`response_format` fields.",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transcription in OpenAI JSON shape (e.g. `{\"text\": …}`)."
          },
          "400": {
            "description": "Malformed request or missing `file` part."
          },
          "413": {
            "description": "Audio exceeds the configured maximum file size."
          },
          "415": {
            "description": "Unsupported media format."
          },
          "500": {
            "description": "Decode or inference failure."
          },
          "503": {
            "description": "No ASR model loaded."
          }
        }
      }
    },
    "/v1/audio/translations": {
      "post": {
        "tags": [
          "openai-compat"
        ],
        "operationId": "translations",
        "requestBody": {
          "description": "OpenAI-compatible translation request (transcribe + translate to English) — `multipart/form-data` with a `file` part plus optional `model`/`response_format` fields.",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "English translation in OpenAI JSON shape (e.g. `{\"text\": …}`)."
          },
          "400": {
            "description": "Malformed request or missing `file` part."
          },
          "413": {
            "description": "Audio exceeds the configured maximum file size."
          },
          "415": {
            "description": "Unsupported media format."
          },
          "500": {
            "description": "Decode or inference failure."
          },
          "503": {
            "description": "No ASR model loaded."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "The `{ \"detail\": string }` body every [`ApiError`] serializes to. `ApiError`\nitself is private and only implements `IntoResponse`, so it cannot derive\n`ToSchema`; this is the single reusable schema every fallible operation\nreferences for its non-200 error responses (rather than re-inlining the\nshape per handler).",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable error description.",
            "example": "Unsupported Content-Type: text/plain"
          }
        }
      },
...
    },
    "securitySchemes": {
      "engine_token_bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "engine_token_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "engine_token"
      }
    }
  },
  "security": [
    {
      "engine_token_bearer": []
    },
    {
      "engine_token_cookie": []
    }
  ]
}